merge examples for rpi and waveshare sense hat
Since both tutorials do more or less the same thing, lets support them
both with the same container, and two notebooks.

The container using a multi-stage build to minimize file-size, but this
is not mandatory.

Additionally, lgpio is compiled from source to preserve compatibility
with more recent debian versions.

Finally the ads1015 example is updated for compatibility with the
adafruit library.